Further added styling and integration with proxy server
