+++ checkout-service/client/src/Index.jsx
@@ -1,5 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from './components/App.jsx';
+import Checkout from './components/Checkout.jsx';
 
-ReactDOM.render(<App />, document.getElementById('checkout-app'));
+ReactDOM.render(<Checkout />, document.getElementById('checkout'));
