+++ checkout-service/package.json
@@ -17,6 +17,7 @@
   },
   "dependencies": {
     "chai-http": "^4.3.0",
+    "cors": "^2.8.5",
     "express": "^4.17.1",
     "faker": "^5.1.0",
     "mongoose": "^5.11.13",
